# perf: undo regression in `as_dict` performance