--- a/frappe/model/base_document.py
+++ b/frappe/model/base_document.py
@@ -110,6 +110,7 @@ class BaseDocument:
 			"_doc_before_save",
 			"_table_fieldnames",
 			"_reserved_keywords",
+			"_permitted_fieldnames",
 			"dont_update_if_missing",
 		)
 	)
@@ -127,10 +128,21 @@ class BaseDocument:
 
 	@property
 	def meta(self):
-		if not (meta := getattr(self, "_meta", None)):
+		meta = getattr(self, "_meta", None)
+		if meta is None:
 			self._meta = meta = frappe.get_meta(self.doctype)
 
 		return meta
+
+	@property
+	def permitted_fieldnames(self):
+		permitted_fieldnames = getattr(self, "_permitted_fieldnames", None)
+		if permitted_fieldnames is None:
+			self._permitted_fieldnames = permitted_fieldnames = get_permitted_fields(
+				doctype=self.doctype, parenttype=getattr(self, "parenttype", None)
+			)
+
+		return permitted_fieldnames
 
 	def __getstate__(self):
 		"""
@@ -150,6 +162,7 @@ class BaseDocument:
 		"""Remove unpicklable values before pickling"""
 
 		state.pop("_meta", None)
+		state.pop("_permitted_fieldnames", None)
 
 	def update(self, d):
 		"""Update multiple fields of a doctype using a dictionary of key-value pairs.
@@ -313,18 +326,14 @@ class BaseDocument:
 		self, sanitize=True, convert_dates_to_str=False, ignore_nulls=False, ignore_virtual=False
 	) -> _dict:
 		d = _dict()
-		permitted_fields = get_permitted_fields(
-			doctype=self.doctype, parenttype=getattr(self, "parenttype", None)
-		)
+		field_values = self.__dict__
 
 		for fieldname in self.meta.get_valid_columns():
-			field_value = getattr(self, fieldname, None)
-
-			# column is valid, we can use getattr
-			d[fieldname] = field_value
+			value = field_values.get(fieldname)
 
 			# if no need for sanitization and value is None, continue
-			if not sanitize and d[fieldname] is None:
+			if not sanitize and value is None:
+				d[fieldname] = None
 				continue
 
 			df = self.meta.get_field(fieldname)
@@ -332,46 +341,47 @@ class BaseDocument:
 
 			if df:
 				if is_virtual_field:
-					if ignore_virtual or fieldname not in permitted_fields:
-						del d[fieldname]
+					if ignore_virtual or fieldname not in self.permitted_fieldnames:
 						continue
 
-					if d[fieldname] is None and (options := getattr(df, "options", None)):
+					if value is None and (options := getattr(df, "options", None)):
 						from frappe.utils.safe_exec import get_safe_globals
 
-						d[fieldname] = frappe.safe_eval(
+						value = frappe.safe_eval(
 							code=options,
 							eval_globals=get_safe_globals(),
 							eval_locals={"doc": self},
 						)
 
-				if isinstance(d[fieldname], list) and df.fieldtype not in table_fields:
+				if isinstance(value, list) and df.fieldtype not in table_fields:
 					frappe.throw(_("Value for {0} cannot be a list").format(_(df.label)))
 
 				if df.fieldtype == "Check":
-					d[fieldname] = 1 if cint(d[fieldname]) else 0
+					value = 1 if cint(value) else 0
 
-				elif df.fieldtype == "Int" and not isinstance(d[fieldname], int):
-					d[fieldname] = cint(d[fieldname])
+				elif df.fieldtype == "Int" and not isinstance(value, int):
+					value = cint(value)
 
-				elif df.fieldtype == "JSON" and isinstance(d[fieldname], dict):
-					d[fieldname] = json.dumps(d[fieldname], sort_keys=True, indent=4, separators=(",", ": "))
+				elif df.fieldtype == "JSON" and isinstance(value, dict):
+					value = json.dumps(value, sort_keys=True, indent=4, separators=(",", ": "))
 
-				elif df.fieldtype in float_like_fields and not isinstance(d[fieldname], float):
-					d[fieldname] = flt(d[fieldname])
+				elif df.fieldtype in float_like_fields and not isinstance(value, float):
+					value = flt(value)
 
-				elif (df.fieldtype in datetime_fields and d[fieldname] == "") or (
-					getattr(df, "unique", False) and cstr(d[fieldname]).strip() == ""
+				elif (df.fieldtype in datetime_fields and value == "") or (
+					getattr(df, "unique", False) and cstr(value).strip() == ""
 				):
-					d[fieldname] = None
+					value = None
 
 			if convert_dates_to_str and isinstance(
-				d[fieldname], (datetime.datetime, datetime.date, datetime.time, datetime.timedelta)
+				value, (datetime.datetime, datetime.date, datetime.time, datetime.timedelta)
 			):
-				d[fieldname] = str(d[fieldname])
+				value = str(value)
 
-			if ignore_nulls and not is_virtual_field and d[fieldname] is None:
-				del d[fieldname]
+			if ignore_nulls and not is_virtual_field and value is None:
+				continue
+
+			d[fieldname] = value
 
 		return d