--- a/frappe/tests/test_perf.py
+++ b/frappe/tests/test_perf.py
@@ -17,6 +17,7 @@ query. This test can be written like this.
 
 """
 import time
+from unittest.mock import patch
 
 from tenacity import retry, retry_if_exception_type, stop_after_attempt, wait_fixed
 
@@ -53,6 +54,18 @@ class TestPerformance(FrappeTestCase):
 
 		with self.assertQueryCount(0):
 			frappe.get_meta("User")
+
+	def test_permitted_fieldnames(self):
+		frappe.clear_cache()
+
+		doc = frappe.new_doc("Prepared Report")
+		# load permitted fieldnames once
+		doc.permitted_fieldnames
+
+		with patch("frappe.model.base_document.get_permitted_fields") as mocked:
+			doc.as_dict()
+			# get_permitted_fields should not be called again
+			mocked.assert_not_called()
 
 	def test_set_value_query_count(self):
 		frappe.db.set_value("User", "Administrator", "interest", "Nothing")