Original fix: https://github.com/frappe/frappe/pull/16549
Regression introduced in: https://github.com/frappe/frappe/pull/19533

### Changes Made

- doc-level cache for `permitted_fieldnames`, created only if required (only if doc has virtual fields)
- Use `__dict__` instead of `getattr`
- Use variable `value` instead of `d[fieldname]`


### Results

#### Sales Invoice - Before

```py
In [3]: %timeit doc.as_dict()
538 µs ± 31.3 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

#### Sales Invoice - After (~64% better)

```py
In [2]: %timeit doc.as_dict()
195 µs ± 777 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```


#### Prepared Report - Before

```py
In [2]: %timeit doc.as_dict()
36.7 µs ± 183 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

#### Prepared Report - After (~79% better)

```py
In [3]: %timeit doc.as_dict()
7.81 µs ± 70.8 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```